requirements.txt
fastapi[standard]
uvicorn
mysql-connector-python
python-dotenv
pydantic
streamlit
requests
pandas
openpyxl